+++ fiberplane-provider-protocol/fiberplane-provider-bindings/src/export.rs
@@ -53,10 +53,6 @@ pub fn get_config_schema() -> ConfigSchema;
 #[fp_bindgen_support::fp_export_signature]
 pub async fn get_supported_query_types(config: ProviderConfig) -> Vec<SupportedQueryType>;
 
-/// Legacy invoke function.
-#[fp_bindgen_support::fp_export_signature]
-pub async fn invoke(request: LegacyProviderRequest, config: ProviderConfig) -> LegacyProviderResponse;
-
 /// Invokes the provider to perform a data request.
 #[fp_bindgen_support::fp_export_signature]
 pub async fn invoke2(request: ProviderRequest) -> Result<Blob, Error>;