+++ fiberplane-provider-protocol/fiberplane-provider-bindings/src/lib.rs
@@ -1,18 +1,3 @@
-/*!
-# Fiberplane Provider Bindings
-
-> Bindings for building Fiberplane Providers
-
-**_Are you looking for examples of providers? Head over to our
-[Providers Repo](https://github.com/fiberplane/providers)._**
-
-These are the low-level bindings for building Fiberplane Providers. While they
-can be used directly, we advise building against the higher-level
-[Fiberplane PDK](https://github.com/fiberplane/providers/tree/main/fiberplane-pdk)
-instead.
-
-*/
-
 #![allow(unused_imports)]
 #[rustfmt::skip]
 mod export;