+++ CHANGELOG.md
@@ -11,9 +11,22 @@ means that individual crates do not strictly follow _SemVer_ although their
 versioning remains _compatible with_ SemVer, i.e. they will not contain breaking
 changes if the major version hasn't changed.
 
+## [Unreleased]
+
+**Breaking changes!** Some breaking changes occurred between beta versions.
+
+### Changed
+
+- `fiberplane_models`: Replaced `AutoSuggestRequest::from_query_data()` with
+  `AutoSuggestRequest::parse()` for consistency with the PDK.
+
+### Removed
+
+- Support for the legacy provider protocol has been removed.
+
 ## [v1.0.0-beta.1] - 2023-02-14
 
 ### Added
 
-- Add support for pinned views (#336)
+- Add support for pinned views
 - Initial open-source release of `fiberplane`.